Correção Método getBase_damage
Correção no tipo do método getBase_damage( )

+++ Armas_Vampire_Survivors/ax.cpp
@@ -36,7 +36,7 @@ Axe::~Axe( )
 }
 
 
-int Axe::getBase_damage( ) const
+double Axe::getBase_damage( ) const
 {
     return base_damage;
 }
